Update README and reorganize game docs structure
The README file has been modified to improve documentation structure. The list of games has been moved to its own separate document instead of being listed directly on the README. This approach allows for better organization and clarity, especially as more games are added. The document for client description has also been improved for better understanding of the business case. A number of new .md files have also been created for each of the games.